ref(core): Don't start transaction for trpc middleware
